compiled lisa with -O2 and removed all the additional warnings that occured then
found a nasty bug in reducing_set.cpp and fixed it

+++ src/algorithm/partition/partition.cpp
@@ -70,7 +70,7 @@ int main(int argc, char *argv[])
     table->fill(FALSE);
     //initiate with 0 and p1 true
     (*table)[0][0] = TRUE;
-    (*table)[0][((*my_werte->PT)[0][0])] = TRUE;
+    (*table)[0][ (int) ((*my_werte->PT)[0][0])] = TRUE;
 
     //fill iterating
     for(int i=1;i<max_n;i++)
@@ -79,7 +79,7 @@ int main(int argc, char *argv[])
 	if((*table)[i-1][j])
 	  (*table)[i][j] = TRUE;
 	//a partition of size-pj was already possible
-	else if(((*my_werte->PT)[i][0]<=j) && ((*table)[i-1][j-((*my_werte->PT)[i][0])]))
+	else if(((*my_werte->PT)[i][0]<=j) && ((*table)[i-1][ (int) (j-(*my_werte->PT)[i][0]) ]))
 	  (*table)[i][j] = TRUE;
       }
